Added quiet option to wget when downloading globalblacklist
This fixes an issue where wget sends progress to stderr causing issues with wrapper scripts which detect the presence of stderr and produce exit status != 1... Since wget progress is not an error, we shouldn't have update-apacheblocker.sh exemplify as stderr. Furthermore, Knowing the progress of wget in a cron isn't necessary so suppressing output with quiet option causes no objectionable harm.